refactor: split codegen program usage
